Issue #11163: Enforce file size on WhitespaceAround inputs
