+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/whitespace/whitespacearound/InputWhitespaceAroundForLoopsAndArrays.java
@@ -1,0 +1,107 @@
+/*
+WhitespaceAround
+allowEmptyConstructors = (default)false
+allowEmptyMethods = (default)false
+allowEmptyTypes = (default)false
+allowEmptyLoops = (default)false
+allowEmptyLambdas = (default)false
+allowEmptyCatches = (default)false
+allowEmptySwitchBlockStatements = (default)false
+ignoreEnhancedForColon = (default)true
+tokens = (default)ASSIGN, BAND, BAND_ASSIGN, BOR, BOR_ASSIGN, BSR, BSR_ASSIGN, BXOR, \
+         BXOR_ASSIGN, COLON, DIV, DIV_ASSIGN, DO_WHILE, EQUAL, GE, GT, LAMBDA, LAND, \
+         LCURLY, LE, LITERAL_CATCH, LITERAL_DO, LITERAL_ELSE, LITERAL_FINALLY, \
+         LITERAL_FOR, LITERAL_IF, LITERAL_RETURN, LITERAL_SWITCH, LITERAL_SYNCHRONIZED, \
+         LITERAL_TRY, LITERAL_WHILE, LOR, LT, MINUS, MINUS_ASSIGN, MOD, MOD_ASSIGN, \
+         NOT_EQUAL, PLUS, PLUS_ASSIGN, QUESTION, RCURLY, SL, SLIST, SL_ASSIGN, SR, \
+         SR_ASSIGN, STAR, STAR_ASSIGN, LITERAL_ASSERT, TYPE_EXTENSION_AND, LITERAL_WHEN
+
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.whitespace.whitespacearound;
+
+/**
+ * Avoid Whitespace violations in for loop.
+ * @author lkuehne
+ * @version 1.0
+ */
+class InputWhitespaceAroundForLoopsAndArrays
+{
+    void forIterator()
+    {
+        // avoid conflict between WhiteSpaceAfter ';' and ParenPad(nospace)
+        for (int i = 0; i++ < 5;) {
+        //                  ^ no whitespace
+    }
+
+        // bug 895072
+    // avoid conflict between ParenPad(space) and NoWhiteSpace before ';'
+    int i = 0;
+    for ( ; i < 5; i++ ) {
+    //   ^ whitespace
+    }
+        for (int anInt : getSomeInts()) {
+            //Should be ignored
+        }
+    }
+
+    int[] getSomeInts() {
+        int i = (int) ( 2 / 3 );
+        return null;
+    }
+
+    public void myMethod() {
+        new Thread() {
+            public void run() {
+            }
+        }.start();
+    }
+
+    public void foo(java.util.List<? extends String[]> bar, Comparable<? super Object[]> baz) { }
+
+    public void mySuperMethod() {
+        Runnable[] runs = new Runnable[] {new Runnable() {
+                public void run() {
+                }
+            },
+            new Runnable() {
+                public void run() {
+                }
+            }}; // violation ''}' is not followed by whitespace'
+        runs[0]
+.
+ run()
+;
+    }
+
+    public void testNullSemi() {
+        return ;
+    }
+
+    public void register(Object obj) { }
+
+    public void doSomething(String args[]) {
+        register(boolean[].class);
+        register( args );
+    }
+
+    public void parentheses() {
+        testNullSemi
+(
+)
+;
+    }
+
+    public static void testNoWhitespaceBeforeEllipses(String ... args) {
+    }
+    public String test() {
+        int pc = 0;
+        return ((100000+pc)+"").substring(1);
+        // 4 violations above:
+        //               ''\+' is not followed by whitespace'
+        //               ''\+' is not preceded with whitespace'
+        //                   ''\+' is not followed by whitespace'
+        //                   ''\+' is not preceded with whitespace'
+    }
+}
