+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/whitespace/whitespacearound/InputWhitespaceAroundKeywordsAndOperators.java
@@ -14,7 +14,7 @@ tokens = (default)ASSIGN, BAND, BAND_ASSIGN, BOR, BOR_ASSIGN, BSR, BSR_ASSIGN, B
          LITERAL_FOR, LITERAL_IF, LITERAL_RETURN, LITERAL_SWITCH, LITERAL_SYNCHRONIZED, \
          LITERAL_TRY, LITERAL_WHILE, LOR, LT, MINUS, MINUS_ASSIGN, MOD, MOD_ASSIGN, \
          NOT_EQUAL, PLUS, PLUS_ASSIGN, QUESTION, RCURLY, SL, SLIST, SL_ASSIGN, SR, \
-         SR_ASSIGN, STAR, STAR_ASSIGN, LITERAL_ASSERT, TYPE_EXTENSION_AND
+         SR_ASSIGN, STAR, STAR_ASSIGN, LITERAL_ASSERT, TYPE_EXTENSION_AND, LITERAL_WHEN
 
 
 */
@@ -96,213 +96,5 @@ class InputWhitespaceAroundKeywordsAndOperators
         {
             return 2;
         }
-    }
-
-    /** test casts **/
-    private void testCasts()
-    {
-        Object o = (Object) new Object();
-        o = (Object)o;
-        o = ( Object ) o;
-        o = (Object)
-            o;
-    }
-
-    /** test questions **/
-    private void testQuestions()
-    {
-        boolean b = (1 == 2)?true:false; // 4 violations
-        b = (1==2) ? false : true; // 2 violations
-    }
-
-    /** star test **/
-    private void starTest()
-    {
-        int x = 2 *3* 4; // 2 violations
-    }
-
-    /** boolean test **/
-    private void boolTest()
-    {
-        boolean a = true;
-        boolean x = ! a;
-        int z = ~1 + ~ 2;
-    }
-
-    /** division test **/
-    private void divTest()
-    {
-        int a = 4 % 2;
-        int b = 4% 2; // violation ''%' is not preceded with whitespace'
-        int c = 4 %2; // violation ''%' is not followed by whitespace'
-        int d = 4%2; // 2 violations
-        int e = 4 / 2;
-        int f = 4/ 2; // violation ''/' is not preceded with whitespace'
-        int g = 4 /2; // violation ''/' is not followed by whitespace'
-        int h = 4/2; // 2 violations
-    }
-
-    /** @return dot test **/
-    private java .lang.  String dotTest()
-    {
-        Object o = new java.lang.Object();
-        o.
-            toString();
-        o
-            .toString();
-        o . toString();
-        return o.toString();
-    }
-
-    /** assert statement test */
-    public void assertTest()
-    {
-        assert true;
-
-        assert true : "Whups";
-
-        // evil colons, should be OK
-        assert "OK".equals(null) ? false : true : "Whups";
-
-        // missing WS around assert
-        assert(true); // violation ''assert' is not followed by whitespace'
-
-        // missing WS around colon
-        assert true:"Whups"; // 2 violations
-    }
-
-    /** another check */
-    void donBradman(Runnable aRun)
-    {
-        donBradman(new Runnable() {
-            public void run() {
-            }
-        });
-
-        final Runnable r = new Runnable() {
-            public void run() {
-            }
-        };
-    }
-
-    /** rfe 521323, detect whitespace before ';' */
-    void rfe521323()
-    {
-        doStuff() ;
-        //       ^ whitespace
-        for (int i = 0 ; i < 5; i++) {
-            //        ^ whitespace
-        }
-    }
-
-
-    /** bug 806243 (NoWhitespaceBeforeCheck violation for anonymous inner class) */
-    private int i ;
-    //           ^ whitespace
-    private int i1, i2, i3 ;
-    //                    ^ whitespace
-    private int i4, i5, i6;
-
-    /** bug 806243 (NoWhitespaceBeforeCheck violation for anonymous inner class) */
-    void bug806243()
-    {
-        Object o = new InputWhitespaceAroundKeywordsAndOperators() {
-            private int j ;
-            //           ^ whitespace
-        };
-    }
-
-    void doStuff() {
-    }
-}
-
-/**
- * Bug 806242 (NoWhitespaceBeforeCheck violation with an interface).
- * @author o_sukhodolsky
- * @version 1.0
- */
-interface IFoo
-{
-    void foo() ;
-    //        ^ whitespace
-}
-
-/**
- * Avoid Whitespace violations in for loop.
- * @author lkuehne
- * @version 1.0
- */
-class SpecialCasesInForLoop
-{
-    void forIterator()
-    {
-        // avoid conflict between WhiteSpaceAfter ';' and ParenPad(nospace)
-        for (int i = 0; i++ < 5;) {
-        //                  ^ no whitespace
-    }
-
-        // bug 895072
-    // avoid conflict between ParenPad(space) and NoWhiteSpace before ';'
-    int i = 0;
-    for ( ; i < 5; i++ ) {
-    //   ^ whitespace
-    }
-        for (int anInt : getSomeInts()) {
-            //Should be ignored
-        }
-    }
-
-    int[] getSomeInts() {
-        int i = (int) ( 2 / 3 );
-        return null;
-    }
-
-    public void myMethod() {
-        new Thread() {
-            public void run() {
-            }
-        }.start();
-    }
-
-    public void foo(java.util.List<? extends String[]> bar, Comparable<? super Object[]> baz) { }
-
-    public void mySuperMethod() {
-        Runnable[] runs = new Runnable[] {new Runnable() {
-                public void run() {
-                }
-            },
-            new Runnable() {
-                public void run() {
-                }
-            }}; // violation ''}' is not followed by whitespace'
-        runs[0]
-.
- run()
-;
-    }
-
-    public void testNullSemi() {
-        return ;
-    }
-
-    public void register(Object obj) { }
-
-    public void doSomething(String args[]) {
-        register(boolean[].class);
-        register( args );
-    }
-
-    public void parentheses() {
-        testNullSemi
-(
-)
-;
-    }
-
-    public static void testNoWhitespaceBeforeEllipses(String ... args) {
-    }
-    public String test() {
-        int pc = 0;
-        return ((100000+pc)+"").substring(1); // 4 violations
     }
 }
